% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiPolygons.R
\name{VoronoiPolygons}
\alias{VoronoiPolygons}
\title{Compute vertices of Voronoi diagram tiles.}
\usage{
VoronoiPolygons(sites, cases = NULL, plot.frame = "sites")
}
\arguments{
\item{sites}{Object. Data frame of sites to construct Voronoi diagram with variables "x" and "y".}

\item{cases}{Object. Data frame of secondary source of data: observations, cases, addresses, etc. with variables "x" and "y".}

\item{plot.frame}{Character. "cases" or "sites". Data to define extent or range of Voronoi diagram.}
}
\description{
For use with polygon functions.
}
\examples{
pump.neighborhoods <- VoronoiPolygons(cholera::pumps, cases)
cholera::snowMap()
invisible(lapply(pump.neighborhoods, polygon))

pump.neighborhoods <- VoronoiPolygons(cholera::pumps, cases = cholera::roads, plot.frame = "cases")
cholera::snowMap()
invisible(lapply(pump.neighborhoods, polygon))
}
