% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoiPolygons.R
\name{VoronoiPolygons}
\alias{VoronoiPolygons}
\title{Compute vertices of Voronoi diagram tiles.}
\usage{
VoronoiPolygons(sites, observed.data = NULL, rw = NULL)
}
\arguments{
\item{sites}{Object. Data frame of sites to construct Voronoi diagram with variables "x" and "y".}

\item{observed.data}{Object. Data frame of secondary source of data: observations, cases, addresses, etc. with variables "x" and "y".}

\item{rw}{Numeric. Vector of corners of the rectangular window: xmin, xmax, ymin, ymax. For deldir::deldir().}
}
\description{
For use with polygon functions.
}
\examples{
polygon.vertices <- VoronoiPolygons(cholera::pumps)
cholera::snowMap()
invisible(lapply(polygon.vertices, polygon))

polygon.vertices <- VoronoiPolygons(cholera::pumps, cholera::roads)
cholera::snowMap()
invisible(lapply(polygon.vertices, polygon))
}
